refactor(server): isolate legacy bootstrap helpers
