+++ services/server/src/legacy-baseline.ts
@@ -1,8 +1,21 @@
-import {
-  EMAIL_NORMALIZATION_VERSION,
-  normalizeEmailAddress
-} from "./email-identity.js";
 import type { DatabaseQueryable } from "./database-types.js";
+import {
+  backfillExternalIdentityEmails,
+  backfillSessionProviders,
+  revokeLegacyHostedBearerGrants
+} from "./legacy-backfills.js";
+import {
+  ensureColumn,
+  ensureConstraint,
+  ensureNotNullable,
+  ensureNullable
+} from "./legacy-schema-helpers.js";
+
+export {
+  backfillExternalIdentityEmails,
+  backfillSessionProviders,
+  revokeLegacyHostedBearerGrants
+} from "./legacy-backfills.js";
 
 /**
  * Frozen bootstrap for an empty database and importer for installations that
@@ -982,133 +995,4 @@ export async function bootstrapLegacyBaseline(db: DatabaseQueryable): Promise<vo
     `CREATE INDEX IF NOT EXISTS authorization_collection_offers_lookup_idx
      ON authorization_collection_offers(authorization_id, expires_at)`
   );
-}
-
-export async function backfillSessionProviders(db: DatabaseQueryable): Promise<void> {
-  await db.query(
-    `UPDATE sessions SET provider = 'github'
-     FROM external_identities
-     WHERE sessions.provider = 'session'
-       AND external_identities.user_id = sessions.user_id
-       AND external_identities.provider = 'github'`
-  );
-}
-
-export async function backfillExternalIdentityEmails(
-  db: DatabaseQueryable
-): Promise<void> {
-  const identities = await db.query<{
-    provider: string;
-    subject: string;
-    email: string;
-  }>(
-    `SELECT provider, subject, email FROM external_identities
-     WHERE email_verified = true
-       AND email IS NOT NULL
-       AND COALESCE(normalized_email, '') = ''`
-  );
-  for (const identity of identities.rows) {
-    let normalizedEmail: string;
-    try {
-      normalizedEmail = normalizeEmailAddress(identity.email);
-    } catch {
-      // Invalid provider presentation data is not an account-linking key.
-      continue;
-    }
-    await db.query(
-      `UPDATE external_identities
-       SET normalized_email = $3, email_normalization_version = $4
-       WHERE provider = $1 AND subject = $2
-         AND COALESCE(normalized_email, '') = ''`,
-      [
-        identity.provider,
-        identity.subject,
-        normalizedEmail,
-        EMAIL_NORMALIZATION_VERSION
-      ]
-    );
-  }
-}
-
-export async function revokeLegacyHostedBearerGrants(db: DatabaseQueryable): Promise<void> {
-  await db.query(
-    `UPDATE grants
-     SET revoked_at = COALESCE(revoked_at, now())
-     WHERE hosted_collection_id IS NOT NULL
-       AND application_origin = 'null'
-       AND proof_public_key IS NULL`
-  );
-  await db.query(
-    `UPDATE access_tokens
-     SET revoked_at = COALESCE(revoked_at, now())
-     WHERE grant_id IN (
-       SELECT id FROM grants
-       WHERE hosted_collection_id IS NOT NULL
-         AND application_origin = 'null'
-         AND proof_public_key IS NULL
-     )`
-  );
-  await db.query(
-    `UPDATE refresh_tokens
-     SET revoked_at = COALESCE(revoked_at, now())
-     WHERE grant_id IN (
-       SELECT id FROM grants
-       WHERE hosted_collection_id IS NOT NULL
-         AND application_origin = 'null'
-         AND proof_public_key IS NULL
-     )`
-  );
-}
-
-async function ensureNullable(db: DatabaseQueryable, table: string, column: string): Promise<void> {
-  const result = await db.query<{ is_nullable: string }>(
-    `SELECT is_nullable FROM information_schema.columns
-     WHERE table_name = $1 AND column_name = $2`,
-    [table, column]
-  );
-  if (result.rows[0]?.is_nullable === "NO") {
-    await db.query(`ALTER TABLE ${table} ALTER COLUMN ${column} DROP NOT NULL`);
-  }
-}
-
-async function ensureNotNullable(
-  db: DatabaseQueryable,
-  table: string,
-  column: string
-): Promise<void> {
-  const result = await db.query<{ is_nullable: string }>(
-    `SELECT is_nullable FROM information_schema.columns
-     WHERE table_name = $1 AND column_name = $2`,
-    [table, column]
-  );
-  if (result.rows[0]?.is_nullable === "YES") {
-    await db.query(`ALTER TABLE ${table} ALTER COLUMN ${column} SET NOT NULL`);
-  }
-}
-
-async function ensureColumn(
-  db: DatabaseQueryable,
-  table: string,
-  column: string,
-  statement: string
-): Promise<void> {
-  const result = await db.query<{ column_name: string }>(
-    `SELECT column_name FROM information_schema.columns WHERE table_name = $1 AND column_name = $2`,
-    [table, column]
-  );
-  if (!result.rows[0]) await db.query(statement);
-}
-
-async function ensureConstraint(
-  db: DatabaseQueryable,
-  table: string,
-  constraint: string,
-  statement: string
-): Promise<void> {
-  const result = await db.query<{ constraint_name: string }>(
-    `SELECT constraint_name FROM information_schema.table_constraints
-     WHERE table_name = $1 AND constraint_name = $2`,
-    [table, constraint]
-  );
-  if (!result.rows[0]) await db.query(statement);
 }
